clean

+++ src/main/java/com/denisimusIT/trecer/AirplaneCharacteristics.java
@@ -1,18 +1,16 @@
 package com.denisimusIT.trecer;
 
 public class AirplaneCharacteristics {
-    private  Long maxSpeed;
-    private  Long maxAcceleration;
-    private  Long speedChangeHeight;
-    private  Long speedChangeCourse;
+    private Long maxSpeed;
+    private Long maxAcceleration;
+    private Long speedChangeHeight;
+    private Long speedChangeCourse;
 
-
-
-    public AirplaneCharacteristics(int maxSpeed, int maxAcceleration, int speedChangeHeight, int speedChangeCourse) {
-        this.maxSpeed = Long.valueOf(maxSpeed);
-        this.maxAcceleration = Long.valueOf(maxAcceleration);
-        this.speedChangeHeight = Long.valueOf(speedChangeHeight);
-        this.speedChangeCourse = Long.valueOf(speedChangeCourse);
+    public AirplaneCharacteristics(Long maxSpeed, Long maxAcceleration, Long speedChangeHeight, Long speedChangeCourse) {
+        this.maxSpeed = maxSpeed;
+        this.maxAcceleration = maxAcceleration;
+        this.speedChangeHeight = speedChangeHeight;
+        this.speedChangeCourse = speedChangeCourse;
     }
 
     public Long getMaxSpeed() {
